ocpb-124 chores: move login to own component folder
