--- frontend/app/login/page.tsx
@@ -1,5 +1,0 @@
-import Login from "@/components/Login";
-
-export default function LoginPage() {
-  return <Login />;
-}
